feat: expose public API in pkg/context for external Go applications
Refactored internal packages to expose core functionality as a public API
that can be imported and used by other Go applications (e.g., serverless
functions, CLI tools, custom automation).

**Breaking Changes:**
- None. Existing Terraform provider functionality unchanged.
- internal/core now re-exports from pkg/context for backward compatibility.

**New Public Package:**
- Created pkg/context with all core functionality:
  - NameGenerator: Generate standardized resource names
  - TagProcessor: Generate cloud-specific tags
  - CloudProvider interface: AWS, Azure, GCP support
  - GitInfo: Repository and commit information
  - Validation functions
  - Tag conversion utilities

**Implementation Details:**
- Moved all core logic from internal/core to pkg/context
- Updated internal/core to re-export from pkg/context (backward compatible)
- Moved all tests to pkg/context (all passing)
- Calculation of name prefix, tags, and data tags works identically

**Example Usage:**
```go
import "github.com/kbrockhoff/terraform-provider-context/pkg/context"

nameGen := &context.NameGenerator{
    Namespace: "myorg",
    Name: "api",
    Environment: "prod",
}
name, _ := nameGen.Generate()
// Result: "myorg-api-prod"
```

**Documentation:**
- Created comprehensive README.md in pkg/context
- Added working example app in examples/client-app
- Updated Makefile to exclude client-app from Terraform tests

**Testing:**
- All existing tests passing (moved to pkg/context)
- Example client app builds and runs successfully
- make test-examples passes for all Terraform configurations

**Files Changed:**
- pkg/context/*.go - New public API
- pkg/context/*_test.go - All tests moved and updated
- pkg/context/README.md - Comprehensive documentation
- internal/core/*.go - Now re-export from pkg/context
- examples/client-app/ - Working example Go application
- Makefile - Exclude client-app from Terraform test loop

Closes beads-mcp-3
