+++ examples/client-app/go.mod
@@ -1,0 +1,7 @@
+module example-client-app
+
+go 1.25.1
+
+require github.com/kbrockhoff/terraform-provider-context v0.0.0
+
+replace github.com/kbrockhoff/terraform-provider-context => ../..
