+++ pkg/context/naming_test.go
@@ -1,4 +1,4 @@
-package core
+package context
 
 import (
 	"testing"
